Refactor date and time
