+++ src/main/java/seedu/address/logic/parser/AddAssignmentCommandParser.java
@@ -8,7 +8,7 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.logic.commands.AddAssignmentCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.student.Assignment;
+import seedu.address.model.assignment.Assignment;
 
 /**
  * Parses input arguments and creates a new AddAssignmentCommand object
