+++ src/main/java/seedu/address/logic/parser/AddLessonCommandParser.java
@@ -10,9 +10,9 @@ import java.util.stream.Stream;
 
 import seedu.address.logic.commands.AddLessonCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.lesson.Date;
+import seedu.address.model.datetimeutil.Date;
 import seedu.address.model.lesson.Lesson;
-import seedu.address.model.lesson.Time;
+import seedu.address.model.datetimeutil.Time;
 import seedu.address.model.student.Email;
 import seedu.address.model.student.Name;
 import seedu.address.model.student.Phone;
