+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -9,8 +9,8 @@ import java.util.Set;
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.StringUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.lesson.Date;
-import seedu.address.model.lesson.Time;
+import seedu.address.model.datetimeutil.Date;
+import seedu.address.model.datetimeutil.Time;
 import seedu.address.model.student.Address;
 import seedu.address.model.student.Email;
 import seedu.address.model.student.Name;
