+++ src/main/java/seedu/address/model/assignment/Assignment.java
@@ -1,4 +1,4 @@
-package seedu.address.model.student;
+package seedu.address.model.assignment;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
