+++ src/main/java/seedu/address/model/student/Student.java
@@ -9,6 +9,7 @@ import java.util.Objects;
 import java.util.Set;
 
 import seedu.address.commons.util.ToStringBuilder;
+import seedu.address.model.assignment.Assignment;
 import seedu.address.model.tag.Tag;
 
 /**
