+++ src/main/java/seedu/address/storage/JsonAdaptedAssignment.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.model.student.Assignment;
+import seedu.address.model.assignment.Assignment;
 
 /**
  * Jackson-friendly version of {@link Assignment}.
