+++ src/main/java/seedu/address/storage/JsonAdaptedLesson.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.model.lesson.Date;
+import seedu.address.model.datetimeutil.Date;
 import seedu.address.model.lesson.Lesson;
-import seedu.address.model.lesson.Time;
+import seedu.address.model.datetimeutil.Time;
 import seedu.address.model.student.Email;
 import seedu.address.model.student.Name;
 import seedu.address.model.student.Subject;
