+++ src/test/java/seedu/address/logic/commands/AddAssignmentCommandTest.java
@@ -10,7 +10,7 @@ import seedu.address.logic.Messages;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.student.Assignment;
+import seedu.address.model.assignment.Assignment;
 import seedu.address.model.student.Student;
 import seedu.address.testutil.StudentBuilder;
