+++ src/test/java/seedu/address/model/lesson/DateTest.java
@@ -9,6 +9,8 @@ import java.time.format.DateTimeFormatter;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.datetimeutil.Date;
+
 
 class DateTest {
