+++ src/test/java/seedu/address/model/lesson/TimeTest.java
@@ -6,6 +6,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.datetimeutil.Time;
+
 class TimeTest {
 
     @Test
